Changed viabilityCharting and a submit from Matt: Hi,
I've attached the updated AlleleIndexer class.

In the doSangerAlleleMapping() method, it was setting
geneLatestMouseStatus instead of mouseStatus, which was the only
difference I could find between that and the javascript in the DIH - my
mistake when copying the code, I think.

I think you only need to run that method once, once all of the data has
been gathered. I've commented out the first call to it, but obviously if
it doesn't work then uncomment.

I also changed "if (mouseStatus != null)" to "if
(!"".equals(mouseStatus))" (lines 381 and 409), which more exactly
matches the Javascript. I doubt it's likely to have a major effect.

Thanks,

Matt